chore: Update project structure by renaming package paths
