--- src/main/java/team/nahyunuk/gsmcertificationsystem/Gsmcertificationsystemv1Application.java
@@ -1,13 +1,0 @@
-package team.nahyunuk.gsmcertificationsystemv1;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Gsmcertificationsystemv1Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Gsmcertificationsystemv1Application.class, args);
-	}
-
-}
