--- src/main/java/team/nahyunuk/gsmcertificationsystem/domain/auth/entity/RefreshToken.java
@@ -1,27 +1,0 @@
-package team.nahyunuk.gsmcertificationsystemv1.domain.auth.entity;
-
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.data.redis.core.RedisHash;
-import org.springframework.data.redis.core.TimeToLive;
-import org.springframework.data.redis.core.index.Indexed;
-
-
-@Getter
-@Builder
-@AllArgsConstructor
-@RedisHash(value = "refreshToken")
-public class RefreshToken {
-
-    @Id
-    @Indexed
-    private Long id;
-
-    @Indexed
-    private String token;
-
-    @TimeToLive
-    private int expTime;
-}
