--- src/main/java/team/nahyunuk/gsmcertificationsystem/domain/user/type/Authority.java
@@ -1,5 +1,0 @@
-package team.nahyunuk.gsmcertificationsystemv1.domain.user.type;
-
-public enum Authority {
-    STUDENT, ADMIN, TEACHER
-}
