--- src/main/java/team/nahyunuk/gsmcertificationsystem/global/redis/util/RedisUtil.java
@@ -1,50 +1,0 @@
-package team.nahyunuk.gsmcertificationsystemv1.global.redis.util;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
-import org.springframework.stereotype.Component;
-
-import java.util.concurrent.TimeUnit;
-
-@Component
-@RequiredArgsConstructor
-public class RedisUtil {
-    private final RedisTemplate<String, Object> redisTemplate;
-    private final RedisTemplate<String, Object> redisBlackListTemplate;
-
-    public Object get(String key) {
-        return redisTemplate.opsForValue().get(key);
-    }
-
-    public void set(String key, Object value, int minutes) {
-        redisTemplate.setValueSerializer(new Jackson2JsonRedisSerializer<>(value.getClass()));
-        redisTemplate.opsForValue().set(key, value, minutes, TimeUnit.MINUTES);
-    }
-
-    public boolean delete(String key) {
-        return redisTemplate.delete(key);
-    }
-
-    public boolean hasKey(String key) {
-        return redisTemplate.hasKey(key);
-    }
-
-    public Object getBlackList(String key) {
-        return redisBlackListTemplate.opsForValue().get(key);
-    }
-
-    public void setBlackList(String key, Object value, Long milliSeconds) {
-        redisBlackListTemplate.setValueSerializer(new Jackson2JsonRedisSerializer<>(value.getClass()));
-        redisBlackListTemplate.opsForValue().set(key, value, milliSeconds, TimeUnit.MILLISECONDS);
-    }
-
-    public boolean deleteBlackList(String key) {
-        return redisBlackListTemplate.delete(key);
-    }
-
-    public boolean hasKayBlackList(String key) {
-        return redisBlackListTemplate.hasKey(key);
-    }
-
-}
