--- src/main/java/team/nahyunuk/gsmcertificationsystem/global/security/auth/CustomUserDetails.java
@@ -1,37 +1,0 @@
-package team.nahyunuk.gsmcertificationsystemv1.global.security.auth;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import team.nahyunuk.gsmcertificationsystemv1.domain.user.entity.User;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-@RequiredArgsConstructor
-public class CustomUserDetails implements UserDetails {
-    private final User user;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        String authority = user.getAuthority().name();
-
-        SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(authority);
-
-        Collection<GrantedAuthority> authorities = new ArrayList<>();
-        authorities.add(simpleGrantedAuthority);
-        return authorities;
-    }
-
-    @Override
-    public String getUsername() {
-        return user.getUsername();
-    }
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-}
