--- src/main/java/team/nahyunuk/gsmcertificationsystem/global/security/handler/CustomAccessDeniedHandler.java
@@ -1,21 +1,0 @@
-package team.nahyunuk.gsmcertificationsystemv1.global.security.handler;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-@Slf4j
-public class CustomAccessDeniedHandler implements AccessDeniedHandler {
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException {
-        log.info("====== ACCESS DENIED ======");
-        response.sendError(HttpServletResponse.SC_FORBIDDEN);
-    }
-}
