+++ src/main/java/team/nahyunuk/gsmcertificationsystem/v1/Gsmcertificationsystemv1Application.java
@@ -1,0 +1,13 @@
+package team.nahyunuk.gsmcertificationsystem;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Gsmcertificationsystemv1Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Gsmcertificationsystemv1Application.class, args);
+	}
+
+}
