+++ src/main/java/team/nahyunuk/gsmcertificationsystem/v1/domain/auth/entity/RefreshToken.java
@@ -1,0 +1,27 @@
+package team.nahyunuk.gsmcertificationsystem.domain.auth.entity;
+
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.data.redis.core.RedisHash;
+import org.springframework.data.redis.core.TimeToLive;
+import org.springframework.data.redis.core.index.Indexed;
+
+
+@Getter
+@Builder
+@AllArgsConstructor
+@RedisHash(value = "refreshToken")
+public class RefreshToken {
+
+    @Id
+    @Indexed
+    private Long id;
+
+    @Indexed
+    private String token;
+
+    @TimeToLive
+    private int expTime;
+}
