+++ src/main/java/team/nahyunuk/gsmcertificationsystem/v1/domain/auth/repository/RefreshTokenRepository.java
@@ -1,0 +1,10 @@
+package team.nahyunuk.gsmcertificationsystem.domain.auth.repository;
+
+import org.springframework.data.repository.CrudRepository;
+import team.nahyunuk.gsmcertificationsystem.domain.auth.entity.RefreshToken;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends CrudRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByToken(String token);
+}
