+++ src/main/java/team/nahyunuk/gsmcertificationsystem/v1/domain/user/entity/User.java
@@ -1,0 +1,32 @@
+package team.nahyunuk.gsmcertificationsystem.domain.user.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.*;
+import team.nahyunuk.gsmcertificationsystem.domain.user.type.Authority;
+
+import java.util.UUID;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private UUID userId;
+
+    private String email;
+    private String username;
+    private String password;
+
+    private int grade;
+    private int classNumber;
+    private int studentNumber;
+
+    private Authority authority;
+}
