+++ src/main/java/team/nahyunuk/gsmcertificationsystem/v1/domain/user/type/Authority.java
@@ -1,0 +1,5 @@
+package team.nahyunuk.gsmcertificationsystem.domain.user.type;
+
+public enum Authority {
+    STUDENT, ADMIN, TEACHER
+}
