+++ src/main/java/team/nahyunuk/gsmcertificationsystem/v1/global/common/response/CommonApiResponse.java
@@ -1,0 +1,17 @@
+package team.nahyunuk.gsmcertificationsystem.global.common.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class CommonApiResponse<T> {
+    private HttpStatus status;
+    private int code;
+    private String message;
+    private T data;
+
+}
