+++ src/main/java/team/nahyunuk/gsmcertificationsystem/v1/global/security/handler/CustomAuthenticationEntryPoint.java
@@ -1,0 +1,20 @@
+package team.nahyunuk.gsmcertificationsystem.v1.global.security.handler;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+@Slf4j
+public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+    @Override
+    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException {
+        log.info("====== AUTHENTICATION DENIED ======");
+        response.sendError(HttpServletResponse.SC_UNAUTHORIZED);
+    }
