+++ src/main/java/team/nahyunuk/gsmcertificationsystem/v1/global/security/jwt/JwtProvider.java
@@ -1,0 +1,126 @@
+package team.nahyunuk.gsmcertificationsystem.v1.global.security.jwt;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import jakarta.annotation.PostConstruct;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
+import team.nahyunuk.gsmcertificationsystem.v1.domain.user.type.Authority;
+import team.nahyunuk.gsmcertificationsystem.v1.global.exception.CustomException;
+import team.nahyunuk.gsmcertificationsystem.v1.global.exception.error.ErrorCode;
+import team.nahyunuk.gsmcertificationsystem.v1.global.redis.util.RedisUtil;
+import team.nahyunuk.gsmcertificationsystem.v1.global.security.auth.CustomUserDetailsService;
+import team.nahyunuk.gsmcertificationsystem.v1.global.security.jwt.dto.TokenDto;
+
+import java.security.Key;
+import java.time.LocalDateTime;
+import java.util.Date;
+import java.util.UUID;
+
+import static team.nahyunuk.gsmcertificationsystem.v1.global.security.filter.JwtFilter.AUTHORIZATION_HEADER;
+import static team.nahyunuk.gsmcertificationsystem.v1.global.security.filter.JwtFilter.BEARER_PREFIX;
+
+@Component
+@RequiredArgsConstructor
+public class JwtProvider {
+    private static final String AUTHORITIES_KEY = "auth";
+    private static final String BEARER_TYPE = "Bearer ";
+    private static final long ACCESS_TOKEN_TIME = 60L * 30 * 4;
+    private static final long REFRESH_TOKEN_TIME = 60L * 60 * 24 * 7;
+
+    @Value("${jwt.secret}")
+    private String secretKey;
+    private static Key key;
+    private final CustomUserDetailsService customUserDetailsService;
+    private final RedisUtil redisUtil;
+
+    @PostConstruct
+    public void init() {
+        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
+        key = Keys.hmacShaKeyFor(keyBytes);
+    }
+
+    public TokenDto generateTokenDto(UUID userId, Authority authority) {
+        return TokenDto.builder()
+                .accessToken(generateAccessToken(userId, authority))
+                .refreshToken(generateRefreshToken(userId))
+                .accessTokenExp(LocalDateTime.now().plusSeconds(ACCESS_TOKEN_TIME))
+                .refreshTokenExp(LocalDateTime.now().plusSeconds(REFRESH_TOKEN_TIME))
+                .build();
+    }
+
+    private String generateAccessToken(UUID userId, Authority authority) {
+        Date accessTokenExp = new Date(System.currentTimeMillis() + ACCESS_TOKEN_TIME * 1000);
+
+        return Jwts.builder()
+                .setSubject(userId.toString())
+                .claim(AUTHORITIES_KEY, authority)
+                .setIssuedAt(new Date())
+                .setExpiration(accessTokenExp)
+                .signWith(key, SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    private String generateRefreshToken(UUID userId) {
+        Date refreshTokenExp = new Date(System.currentTimeMillis() + REFRESH_TOKEN_TIME * 1000);
+
+        return Jwts.builder()
+                .setSubject(userId.toString())
+                .setExpiration(refreshTokenExp)
+                .signWith(key, SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    public boolean validateToken(String token) {
+        try {
+            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
+            return !redisUtil.hasKayBlackList(token);
+        } catch (ExpiredJwtException e) {
+            throw new CustomException(ErrorCode.EXPIRED_TOKEN);
+        } catch (Exception e) {
+            throw new CustomException(ErrorCode.INVALID_TOKEN);
+        }
+    }
+
+    public Authentication getAuthentication(String accessToken) {
+        Claims claims = parseClaims(accessToken);
+
+        if (claims.get(AUTHORITIES_KEY) == null) {
+            throw new CustomException(ErrorCode.INVALID_TOKEN);
+        }
+
+        UserDetails principal = customUserDetailsService.loadUserByUsername(claims.getSubject());
+        return new UsernamePasswordAuthenticationToken(principal, "", principal.getAuthorities());
+    }
+
+    private Claims parseClaims(String accessToken) {
+        try {
+            return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(accessToken).getBody();
+        } catch (ExpiredJwtException e) {
+            return e.getClaims();
+        }
+    }
+
+    public String resolveToken(HttpServletRequest request) {
+        String bearerToken = request.getHeader(AUTHORIZATION_HEADER);
+        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith(BEARER_PREFIX)) {
+            return bearerToken.substring(7);
+        }
+        return null;
+    }
+
+    public String parseRefreshToken(String refreshToken) {
+        if (refreshToken.startsWith(BEARER_TYPE)) {
+            return refreshToken.replace(BEARER_TYPE, "");
+        } else {
+            return n
